% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{fisher}
\alias{fisher}
\title{Fisher combining function}
\usage{
fisher(pvalues)
}
\arguments{
\item{pvalues}{Array of p-values}
}
\value{
Combined p-value using fisher's method
}
\description{
This function takes an array of p-values and returns a combined p-value using fisher's combining function:
\eqn{-2 \sum_i \log(p_i)}
}
\examples{
fisher(pvalues = c(.05, .1, .5))

}
