% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{liptak}
\alias{liptak}
\title{Liptak combining function}
\usage{
liptak(pvalues)
}
\arguments{
\item{pvalues}{Array of p-values}
}
\value{
Combined p-value using Liptak's method
}
\description{
This function takes an array of p-values and returns a combined p-value using Liptak's combining function:
\eqn{\sum_i \Phi^{-1}(1-p_i)} where \eqn{\Phi} is the CDF of the Normal distribution
}
\examples{
liptak(pvalues = c(.05, .1, .5))

}
