% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_stats.R
\name{one_way_anova_stat}
\alias{one_way_anova_stat}
\title{Calculate one-way anova test statistic}
\usage{
one_way_anova_stat(df, group_col, outcome_col)
}
\arguments{
\item{df}{A data frame}

\item{group_col}{The name of the column in df that corresponds to the group label}

\item{outcome_col}{The name of the column in df that corresponds to the outcome variable}
}
\value{
The one-way anova test statistic:
\eqn{\sum_{g=1}^G n_g(\overline{X_g} - \overline{X})^2} where \eqn{g} indexes the groups
}
\description{
This function takes a data frame, and group and outcome column names as input
and returns the one-way anova test statistic
}
