% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{tippett}
\alias{tippett}
\title{Tippett combining function}
\usage{
tippett(pvalues)
}
\arguments{
\item{pvalues}{Array of p-values}
}
\value{
Combined p-value using Tippett's method
}
\description{
This function takes an array of p-values and returns a combined p-value using Tippett's combining function:
\eqn{\max_i \{1-p_i\}}
}
\examples{
tippett(pvalues = c(.05, .1, .5))

}
