% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_stats.R
\name{ttest_stat}
\alias{ttest_stat}
\title{Calculate t-test statistic}
\usage{
ttest_stat(df, group_col, outcome_col)
}
\arguments{
\item{df}{A data frame}

\item{group_col}{The name of the column in df that corresponds to the group label}

\item{outcome_col}{The name of the column in df that corresponds to the outcome variable}
}
\value{
The t test statistic
}
\description{
This function takes a data frame, and group and outcome column names as input
and returns the t test statistic
}
